test: refatorando e adicionando testes unitários para a classe Database

+++ src/database/index.ts
@@ -3,23 +3,40 @@ import knex, { Knex } from 'knex';
 import config from '../../config';
 
 class Database {
-	public readonly client: Knex;
+	private client: Knex | null;
 
 	constructor() {
-		this.client = knex({
-			client: 'pg',
-			connection: {
-				host: config.postgresHost,
-				port: config.postgresPort,
-				user: config.postgresUser,
-				password: config.postgresPassword,
-				database: config.postgresDb,
-			},
-		});
+		this.client = null;
+	}
+
+	async connect() {
+		if (!this.client) {
+			this.client = knex({
+				client: 'pg',
+				connection: {
+					host: config.postgresHost,
+					port: config.postgresPort,
+					user: config.postgresUser,
+					password: config.postgresPassword,
+					database: config.postgresDb,
+				},
+			});
+		}
 	}
 
 	async disconnect() {
-		await this.client.destroy();
+		if (this.client) {
+			await this.client.destroy();
+			this.client = null;
+		}
+	}
+
+	getClient() {
+		if (!this.client) {
+			throw new Error('Database not connected!');
+		}
+
+		return this.client;
 	}
 }
